+++ CHANGELOG.md
@@ -6,11 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- Configuration property for dry heracleum color.
+
+### Changed
+
+- Dry variants of heracleum plant now can't stay in air, it uses logic from non-dry version.
 
 ## [0.0.0-alpha] - 2022-07-24
 ### Added
 
-- Mod now requries [MJUtils 1.6+](https://github.com/MJaroslav/MJUtils).
+- Mod now requires [MJUtils 1.6+](https://github.com/MJaroslav/MJUtils).
 - Heracleum plant block in 16 variants: *dry/not-dry* X *small/stem/middle-stem/top* X *blooming/not-blooming*.
 - Heracleum stem (item) in 2 variants: *dry* and *not dry*, can place stem block.
 - Heracleum umbels (item) can place sprout block.